Draft for NDBC netcdf reader, move existing into ndbc_ascii
